move `App` to package `internal/app`
